requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cycler==0.11.0
Django==3.1.7
filterpy==1.4.5
Flask==2.0.2
flatbuffers==1.12
fonttools==4.29.0
gast==0.3.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
imageio==2.14.1
itsdangerous==2.0.1
Jinja2==3.0.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
key-generator==1.0.3
kiwisolver==1.3.2
#lap==0.4.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
munkres==1.1.4
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.3.4
Pillow==9.0.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tifffile==2021.11.2
tk==0.1.0
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.12.1
yapf==0.32.0
